MAINT: Refactor both features into a base feature
This should reduce code duplication and make adding new features easier
